Convert size fields to uint32

+++ src/conn_mgmt.go
@@ -53,11 +53,11 @@ func handle_conn_send_data(msg message) message {
 
     conn := ConnectionTable[connSendData.conn_id]
 
-    if connSendData.seq_no + connSendData.size != conn.sendSeqNo + connSendData.size {
+    if connSendData.seq_no + uint32(connSendData.size) != conn.sendSeqNo + uint32(connSendData.size) {
         return msg_checksum_invalid{}
     }
 
-    connSendData.seq_no = connSendData.seq_no + connSendData.size
+    connSendData.seq_no = connSendData.seq_no + uint32(connSendData.size)
     if connSendData.seq_no > 2**32 - 1 {
         connSendData.seq_no = connSendData - 2**32
     }
